Revert "Move verification data to the public package"
This reverts commit adaa439cd3c9edde4043df1f2745fb2c699c3fea.
