+++ src/main/java/org/mockito/internal/verification/InOrderWrapper.java
@@ -6,7 +6,7 @@ package org.mockito.internal.verification;
 
 import org.mockito.internal.InOrderImpl;
 import org.mockito.internal.invocation.finder.VerifiableInvocationsFinder;
-import org.mockito.verification.VerificationData;
+import org.mockito.internal.verification.api.VerificationData;
 import org.mockito.internal.verification.api.VerificationDataInOrderImpl;
 import org.mockito.internal.verification.api.VerificationInOrderMode;
 import org.mockito.invocation.Invocation;
