+++ src/main/java/org/mockito/internal/verification/Times.java
@@ -10,7 +10,7 @@ import static org.mockito.internal.verification.checkers.MissingInvocationChecke
 import java.util.List;
 import org.mockito.exceptions.base.MockitoException;
 import org.mockito.internal.invocation.InvocationMatcher;
-import org.mockito.verification.VerificationData;
+import org.mockito.internal.verification.api.VerificationData;
 import org.mockito.internal.verification.api.VerificationDataInOrder;
 import org.mockito.internal.verification.api.VerificationInOrderMode;
 import org.mockito.internal.verification.checkers.NumberOfInvocationsChecker;
