+++ src/main/java/org/mockito/internal/verification/VerificationOverTimeImpl.java
@@ -6,7 +6,7 @@ package org.mockito.internal.verification;
 
 import org.mockito.exceptions.base.MockitoAssertionError;
 import org.mockito.internal.util.Timer;
-import org.mockito.verification.VerificationData;
+import org.mockito.internal.verification.api.VerificationData;
 import org.mockito.verification.VerificationMode;
 
 /**
