+++ src/main/java/org/mockito/internal/verification/api/VerificationData.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2007 Mockito contributors
  * This program is made available under the terms of the MIT License.
  */
-package org.mockito.verification;
+package org.mockito.internal.verification.api;
 
 import java.util.List;
