+++ src/main/java/org/mockito/verification/VerificationMode.java
@@ -6,6 +6,7 @@
 package org.mockito.verification;
 
 import org.mockito.Mockito;
+import org.mockito.internal.verification.api.VerificationData;
 
 /**
  * Allows verifying that certain behavior happened at least once / exact number
