+++ src/test/java/org/mockitousage/debugging/VerificationListenerCallBackTest.java
@@ -5,7 +5,7 @@ import org.mockito.exceptions.base.MockitoAssertionError;
 import org.mockito.exceptions.misusing.NullInsteadOfMockException;
 import org.mockito.internal.progress.MockingProgress;
 import org.mockito.internal.progress.ThreadSafeMockingProgress;
-import org.mockito.verification.VerificationData;
+import org.mockito.internal.verification.api.VerificationData;
 import org.mockito.listeners.VerificationListener;
 import org.mockito.verification.VerificationMode;
 import org.mockito.verification.VerificationSucceededEvent;
